api_telegram_rossmann/requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==2.0.0
idna==2.8
importlib-metadata==1.5.0
inflection==0.5.1
itsdangerous==2.0.0
jedi==0.15.2
Jinja2==3.0.0
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.0
matplotlib==3.4.2
mistune==0.8.4
more-itertools==4.2.0
numpy==1.20.3
pandas==1.2.4
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
Pygments==2.3.1
pyparsing==2.4.7
pyrsistent==0.15.5
python-dateutil==2.7.3
python-telegram-bot==13.7
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.22.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.14.0
tabulate==0.8.9
terminado==0.8.2
testpath==0.4.4
tornado==5.1.1
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.0.0
widgetsnbextension==2.0.0
zipp==1.0.0
